+++ src/Symfony/Component/HttpKernel/Controller/ArgumentResolverInterface.php
@@ -24,8 +24,8 @@ interface ArgumentResolverInterface
     /**
      * Returns the arguments to pass to the controller.
      *
-     * @param Request  $request    A Request instance
-     * @param callable $controller A PHP callable
+     * @param Request  $request
+     * @param callable $controller
      *
      * @return array An array of arguments to pass to the controller
      *