+++ src/Symfony/Component/HttpKernel/Controller/ArgumentValueResolverInterface.php
@@ -1,0 +1,45 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\HttpKernel\Controller;
+
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpKernel\ControllerMetadata\ArgumentMetadata;
+
+/**
+ * Responsible for resolving the value of an argument based on its metadata.
+ *
+ * @author Iltar van der Berg <kjarli@gmail.com>
+ */
+interface ArgumentValueResolverInterface
+{
+    /**
+     * Whether this resolver can resolve can resolve the value for the given ArgumentMetadata.
+     *
+     * @param Request          $request
+     * @param ArgumentMetadata $argument
+     *
+     * @return bool
+     */
+    public function supports(Request $request, ArgumentMetadata $argument);
+
+    /**
+     * Yield the possible value(s).
+     *
+     * An implementation must yield at least one value.
+     *
+     * @param Request          $request
+     * @param ArgumentMetadata $argument
+     *
+     * @return \Generator
+     */
+    public function resolve(Request $request, ArgumentMetadata $argument);
+}