+++ src/Symfony/Component/HttpKernel/HttpKernel.php
@@ -11,7 +11,6 @@
 
 namespace Symfony\Component\HttpKernel;
 
-use Symfony\Component\HttpKernel\Controller\ArgumentResolver;
 use Symfony\Component\HttpKernel\Controller\ArgumentResolverInterface;
 use Symfony\Component\HttpKernel\Controller\ControllerResolverInterface;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
@@ -40,14 +39,6 @@ class HttpKernel implements HttpKernelInterface, TerminableInterface
     protected $requestStack;
     private $argumentResolver;
 
-    /**
-     * Constructor.
-     *
-     * @param EventDispatcherInterface    $dispatcher       An EventDispatcherInterface instance
-     * @param ControllerResolverInterface $resolver         A ControllerResolverInterface instance
-     * @param RequestStack                $requestStack     A stack for master/sub requests
-     * @param ArgumentResolverInterface   $argumentResolver An ArgumentResolverInterface instance
-     */
     public function __construct(EventDispatcherInterface $dispatcher, ControllerResolverInterface $resolver, RequestStack $requestStack = null, ArgumentResolverInterface $argumentResolver = null)
     {
         $this->dispatcher = $dispatcher;