+++ src/Symfony/Component/HttpKernel/Tests/HttpKernelTest.php
@@ -11,6 +11,10 @@
 
 namespace Symfony\Component\HttpKernel\Tests;
 
+use Symfony\Component\EventDispatcher\EventDispatcherInterface;
+use Symfony\Component\HttpFoundation\RequestStack;
+use Symfony\Component\HttpKernel\Controller\ArgumentResolverInterface;
+use Symfony\Component\HttpKernel\Controller\ControllerResolverInterface;
 use Symfony\Component\HttpKernel\HttpKernel;
 use Symfony\Component\HttpKernel\HttpKernelInterface;
 use Symfony\Component\HttpKernel\KernelEvents;
@@ -28,7 +32,7 @@ class HttpKernelTest extends \PHPUnit_Framework_TestCase
      */
     public function testHandleWhenControllerThrowsAnExceptionAndCatchIsTrue()
     {
-        $kernel = new HttpKernel(new EventDispatcher(), $this->getResolver(function () { throw new \RuntimeException(); }));
+        $kernel = $this->getHttpKernel(new EventDispatcher(), function () { throw new \RuntimeException(); });
 
         $kernel->handle(new Request(), HttpKernelInterface::MASTER_REQUEST, true);
     }
@@ -38,7 +42,7 @@ class HttpKernelTest extends \PHPUnit_Framework_TestCase
      */
     public function testHandleWhenControllerThrowsAnExceptionAndCatchIsFalseAndNoListenerIsRegistered()
     {
-        $kernel = new HttpKernel(new EventDispatcher(), $this->getResolver(function () { throw new \RuntimeException(); }));
+        $kernel = $this->getHttpKernel(new EventDispatcher(), function () { throw new \RuntimeException(); });
 
         $kernel->handle(new Request(), HttpKernelInterface::MASTER_REQUEST, false);
     }
@@ -50,7 +54,7 @@ class HttpKernelTest extends \PHPUnit_Framework_TestCase
             $event->setResponse(new Response($event->getException()->getMessage()));
         });
 
-        $kernel = new HttpKernel($dispatcher, $this->getResolver(function () { throw new \RuntimeException('foo'); }));
+        $kernel = $this->getHttpKernel($dispatcher, function () { throw new \RuntimeException('foo'); });
         $response = $kernel->handle(new Request(), HttpKernelInterface::MASTER_REQUEST, true);
 
         $this->assertEquals('500', $response->getStatusCode());
@@ -66,7 +70,7 @@ class HttpKernelTest extends \PHPUnit_Framework_TestCase
             // should set a response, but does not
         });
 
-        $kernel = new HttpKernel($dispatcher, $this->getResolver(function () use ($exception) { throw $exception; }));
+        $kernel = $this->getHttpKernel($dispatcher, function () use ($exception) { throw $exception; });
 
         try {
             $kernel->handle(new Request(), HttpKernelInterface::MASTER_REQUEST, true);
@@ -83,7 +87,7 @@ class HttpKernelTest extends \PHPUnit_Framework_TestCase
             $event->setResponse(new RedirectResponse('/login', 301));
         });
 
-        $kernel = new HttpKernel($dispatcher, $this->getResolver(function () { throw new AccessDeniedHttpException(); }));
+        $kernel = $this->getHttpKernel($dispatcher, function () { throw new AccessDeniedHttpException(); });
         $response = $kernel->handle(new Request());
 
         $this->assertEquals('301', $response->getStatusCode());
@@ -97,7 +101,7 @@ class HttpKernelTest extends \PHPUnit_Framework_TestCase
             $event->setResponse(new Response($event->getException()->getMessage()));
         });
 
-        $kernel = new HttpKernel($dispatcher, $this->getResolver(function () { throw new MethodNotAllowedHttpException(array('POST')); }));
+        $kernel = $this->getHttpKernel($dispatcher, function () { throw new MethodNotAllowedHttpException(array('POST')); });
         $response = $kernel->handle(new Request());
 
         $this->assertEquals('405', $response->getStatusCode());
@@ -114,7 +118,7 @@ class HttpKernelTest extends \PHPUnit_Framework_TestCase
             $event->setResponse(new Response('', $responseStatusCode, array('X-Status-Code' => $expectedStatusCode)));
         });
 
-        $kernel = new HttpKernel($dispatcher, $this->getResolver(function () { throw new \RuntimeException(); }));
+        $kernel = $this->getHttpKernel($dispatcher, function () { throw new \RuntimeException(); });
         $response = $kernel->handle(new Request());
 
         $this->assertEquals($expectedStatusCode, $response->getStatusCode());
@@ -138,7 +142,7 @@ class HttpKernelTest extends \PHPUnit_Framework_TestCase
             $event->setResponse(new Response('hello'));
         });
 
-        $kernel = new HttpKernel($dispatcher, $this->getResolver());
+        $kernel = $this->getHttpKernel($dispatcher);
 
         $this->assertEquals('hello', $kernel->handle(new Request())->getContent());
     }
@@ -149,7 +153,7 @@ class HttpKernelTest extends \PHPUnit_Framework_TestCase
     public function testHandleWhenNoControllerIsFound()
     {
         $dispatcher = new EventDispatcher();
-        $kernel = new HttpKernel($dispatcher, $this->getResolver(false));
+        $kernel = $this->getHttpKernel($dispatcher, false);
 
         $kernel->handle(new Request());
     }
@@ -158,7 +162,7 @@ class HttpKernelTest extends \PHPUnit_Framework_TestCase
     {
         $response = new Response('foo');
         $dispatcher = new EventDispatcher();
-        $kernel = new HttpKernel($dispatcher, $this->getResolver(function () use ($response) { return $response; }));
+        $kernel = $this->getHttpKernel($dispatcher, function () use ($response) { return $response; });
 
         $this->assertSame($response, $kernel->handle(new Request()));
     }
@@ -166,7 +170,7 @@ class HttpKernelTest extends \PHPUnit_Framework_TestCase
     public function testHandleWhenTheControllerIsAnObjectWithInvoke()
     {
         $dispatcher = new EventDispatcher();
-        $kernel = new HttpKernel($dispatcher, $this->getResolver(new Controller()));
+        $kernel = $this->getHttpKernel($dispatcher, new Controller());
 
         $this->assertResponseEquals(new Response('foo'), $kernel->handle(new Request()));
     }
@@ -174,7 +178,7 @@ class HttpKernelTest extends \PHPUnit_Framework_TestCase
     public function testHandleWhenTheControllerIsAFunction()
     {
         $dispatcher = new EventDispatcher();
-        $kernel = new HttpKernel($dispatcher, $this->getResolver('Symfony\Component\HttpKernel\Tests\controller_func'));
+        $kernel = $this->getHttpKernel($dispatcher, 'Symfony\Component\HttpKernel\Tests\controller_func');
 
         $this->assertResponseEquals(new Response('foo'), $kernel->handle(new Request()));
     }
@@ -182,7 +186,7 @@ class HttpKernelTest extends \PHPUnit_Framework_TestCase
     public function testHandleWhenTheControllerIsAnArray()
     {
         $dispatcher = new EventDispatcher();
-        $kernel = new HttpKernel($dispatcher, $this->getResolver(array(new Controller(), 'controller')));
+        $kernel = $this->getHttpKernel($dispatcher, array(new Controller(), 'controller'));
 
         $this->assertResponseEquals(new Response('foo'), $kernel->handle(new Request()));
     }
@@ -190,7 +194,7 @@ class HttpKernelTest extends \PHPUnit_Framework_TestCase
     public function testHandleWhenTheControllerIsAStaticArray()
     {
         $dispatcher = new EventDispatcher();
-        $kernel = new HttpKernel($dispatcher, $this->getResolver(array('Symfony\Component\HttpKernel\Tests\Controller', 'staticcontroller')));
+        $kernel = $this->getHttpKernel($dispatcher, array('Symfony\Component\HttpKernel\Tests\Controller', 'staticcontroller'));
 
         $this->assertResponseEquals(new Response('foo'), $kernel->handle(new Request()));
     }
@@ -201,7 +205,7 @@ class HttpKernelTest extends \PHPUnit_Framework_TestCase
     public function testHandleWhenTheControllerDoesNotReturnAResponse()
     {
         $dispatcher = new EventDispatcher();
-        $kernel = new HttpKernel($dispatcher, $this->getResolver(function () { return 'foo'; }));
+        $kernel = $this->getHttpKernel($dispatcher, function () { return 'foo'; });
 
         $kernel->handle(new Request());
     }
@@ -212,7 +216,8 @@ class HttpKernelTest extends \PHPUnit_Framework_TestCase
         $dispatcher->addListener(KernelEvents::VIEW, function ($event) {
             $event->setResponse(new Response($event->getControllerResult()));
         });
-        $kernel = new HttpKernel($dispatcher, $this->getResolver(function () { return 'foo'; }));
+
+        $kernel = $this->getHttpKernel($dispatcher, function () { return 'foo'; });
 
         $this->assertEquals('foo', $kernel->handle(new Request())->getContent());
     }
@@ -223,7 +228,7 @@ class HttpKernelTest extends \PHPUnit_Framework_TestCase
         $dispatcher->addListener(KernelEvents::RESPONSE, function ($event) {
             $event->setResponse(new Response('foo'));
         });
-        $kernel = new HttpKernel($dispatcher, $this->getResolver());
+        $kernel = $this->getHttpKernel($dispatcher);
 
         $this->assertEquals('foo', $kernel->handle(new Request())->getContent());
     }
@@ -231,7 +236,7 @@ class HttpKernelTest extends \PHPUnit_Framework_TestCase
     public function testTerminate()
     {
         $dispatcher = new EventDispatcher();
-        $kernel = new HttpKernel($dispatcher, $this->getResolver());
+        $kernel = $this->getHttpKernel($dispatcher);
         $dispatcher->addListener(KernelEvents::TERMINATE, function ($event) use (&$called, &$capturedKernel, &$capturedRequest, &$capturedResponse) {
             $called = true;
             $capturedKernel = $event->getKernel();
@@ -255,29 +260,33 @@ class HttpKernelTest extends \PHPUnit_Framework_TestCase
         $stack->expects($this->at(1))->method('pop');
 
         $dispatcher = new EventDispatcher();
-        $kernel = new HttpKernel($dispatcher, $this->getResolver(), $stack);
+        $kernel = $this->getHttpKernel($dispatcher, null, $stack);
 
         $kernel->handle($request, HttpKernelInterface::MASTER_REQUEST);
     }
 
-    protected function getResolver($controller = null)
+    private function getHttpKernel(EventDispatcherInterface $eventDispatcher, $controller = null, RequestStack $requestStack = null)
     {
         if (null === $controller) {
             $controller = function () { return new Response('Hello'); };
         }
 
-        $resolver = $this->getMock('Symfony\\Component\\HttpKernel\\Controller\\ControllerResolverInterface');
-        $resolver->expects($this->any())
+        $controllerResolver = $this->getMock(ControllerResolverInterface::class);
+        $controllerResolver
+            ->expects($this->any())
             ->method('getController')
             ->will($this->returnValue($controller));
-        $resolver->expects($this->any())
+
+        $argumentResolver = $this->getMock(ArgumentResolverInterface::class);
+        $argumentResolver
+            ->expects($this->any())
             ->method('getArguments')
             ->will($this->returnValue(array()));
 
-        return $resolver;
+        return new HttpKernel($eventDispatcher, $controllerResolver, $requestStack, $argumentResolver);
     }
 
-    protected function assertResponseEquals(Response $expected, Response $actual)
+    private function assertResponseEquals(Response $expected, Response $actual)
     {
         $expected->setDate($actual->getDate());
         $this->assertEquals($expected, $actual);