+++ src/Symfony/Component/HttpKernel/CHANGELOG.md
@@ -4,9 +4,10 @@ CHANGELOG
 3.1.0
 -----
  * deprecated passing objects as URI attributes to the ESI and SSI renderers
- * Added an `ArgumentResolver` with `getArguments()` and the respective interface `ArgumentResolverInterface`
- * Deprecated `ControllerResolver::getArguments()`, which uses the `ArgumentResolver` as BC layer by extending it
- * The `HttpKernel` now accepts an additional argument for an `ArgumentResolver`
+ * Added a `LegacyArgumentResolver` with `getArguments()` and the corresponding interface `ArgumentResolverInterface`
+ * Deprecated `ControllerResolver::getArguments()`, which uses the `LegacyArgumentResolver` as BC layer by extending it
+ * The `HttpKernel` now accepts an additional argument for an `ArgumentResolverInterface`
+ * Added the `ArgumentResolver` which features an extension point to resolve arguments in a more dynamic way
 
 3.0.0
 -----